+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.0] - 2025-01-01
 ### Added
 - LerpAsync function
+- InterpolateAsync function
 - WaitWhileAsync function
 - WaitUntilAsync function
 - CHANGELOG